Chore/add testing with pytest

+++ reflash/reflash.py
@@ -6,12 +6,15 @@ import sqlite3
 import threading
 
 class State(object):
-    def __init__(self, db):
-        self.db = db
+    def __init__(self, db_file):
+        self.db = sqlite3.connect(db_file, check_same_thread=False)
         self.db.row_factory = sqlite3.Row
         self.cur = self.db.cursor()
         self.lock = threading.Lock()
 
+        self.settings_darkmode = True
+        self.settings_reboot_when_done = False
+        self.settings_enable_ssh = False
         self.download_state = "IDLE"
         self.download_bytes_now = 0
         self.download_bytes_total = 1
@@ -31,8 +34,16 @@ class State(object):
         self.backup_total = 0
         self.backup_start_time = 0
         self.backup_log = ""
+        self.upload_state = "IDLE"
+        self.upload_filename = ""
+        self.upload_bytes_total = 1
+        self.upload_bytes_now = 0
+        self.upload_start_time = 0
 
         self.fields = [
+            'settings_darkmode',
+            'settings_reboot_when_done',
+            'settings_enable_ssh',
             'download_bytes_now',
             'download_bytes_total',
             'download_state',
@@ -51,8 +62,13 @@ class State(object):
             'backup_filename',
             'backup_total',
             'backup_start_time',
-            'backup_log'
-            ]
+            'backup_log',
+            'upload_state',
+            'upload_filename',
+            'upload_bytes_total',
+            'upload_bytes_now',
+            'upload_start_time',
+        ]
 
     def db_exists(self):
         self.cur.execute("SELECT count(name) FROM sqlite_master WHERE type='table' AND name='state'")
@@ -62,12 +78,10 @@ class State(object):
     def create_default(self):
         self.cur.execute("DROP TABLE IF EXISTS state")
         self.cur.execute("CREATE TABLE state(name, value, type)")
-
         for line in self.fields:
             ins = [line, getattr(self, line), type(getattr(self, line)).__name__]
             self.cur.execute("INSERT INTO state VALUES(?,?,?)", ins)
         self.db.commit()
-
 
     def assign_var(self, name, value, type_name):
         var = __builtins__[type_name](value)
@@ -95,17 +109,30 @@ class Reflash:
     def __init__(self, settings):
         self.reflash_version_file = settings.get("version_file")
         self.images_folder = settings.get("images_folder")
-        self.settings_folder = settings.get("settings_folder")
         self.executor = concurrent.futures.ThreadPoolExecutor(max_workers=4)
-        self.state = State(settings.get("db"))
+        self.sudo = "sudo" if settings.get("use_sudo") else ""
+
+        self.state = State(settings.get("db_file"))
         if not self.state.db_exists():
             self.state.create_default()
         self.state.load()
 
-    def get_reflash_version(self):
-        with open(self.reflash_version_file, "r") as f:
-            version = f.read().replace("\n", "")
-            return version
+    def get_version(self):
+        return self._run_system_command(f"cat {self.reflash_version_file}")
+
+    def refresh(self):
+        self.state.load()
+
+    def get_state(self):
+        if self.state.install_state == "INSTALLING":
+            return "INSTALLING"
+        if self.state.download_state == "DOWNLOADING":
+            return "DOWNLOADING"
+        if self.state.backup_state == "INSTALLING":
+            return "BACKING_UP"
+        if self.state.upload_state == "UPLOADING":
+            return "UPLOADING"
+        return"IDLE"
 
     def check_file_integrity(self, filename):
         path = self.images_folder + "/"+filename +".img.xz"
@@ -125,23 +152,18 @@ class Reflash:
         ]
         return images
 
-    def download_version(self, refactor_image, start_time):
-        url = refactor_image["url"]
+    def download_refactor(self, url, size, filename, start_time):
+        url = url
         self.state.download_bytes_now = 0
         self.state.download_start_time = start_time
         self.state.download_state = "DOWNLOADING"
-        self.state.download_bytes_total = refactor_image["size"]
-        self.state.download_filename = refactor_image["name"]
+        self.state.download_bytes_total = size
+        self.state.download_filename = filename
         self.state.save()
-        self.executor.submit(self.download_refactor, url, refactor_image["name"])
-
-    def save_file_chunk(self, chunk, filename, is_new_file):
-        open_rights = "wb+" if is_new_file else "ab+"
-        with open(self.images_folder+"/"+filename, open_rights) as f:
-            f.write(chunk)
+        self.executor.submit(self._ex_download_refactor, url, filename)
         return True
 
-    def download_refactor(self, url, filename):
+    def _ex_download_refactor(self, url, filename):
         import requests
         # Answer by Dennis Patterson
         # https://stackoverflow.com/questions/53101597/how-to-download-binary-file-using-requests
@@ -161,10 +183,16 @@ class Reflash:
         self.state.save()
 
     def cancel_download(self):
-        os.remove(self.images_folder + "/" + self.state.download_filename)
+        if self.state.download_state != "DOWNLOADING":
+            return False
+        try:
+            os.remove(self.images_folder + "/" + self.state.download_filename)
+        except FileNotFoundError:
+            pass
         self.state.download_state = "CANCELLED"
         self.state.save()
-
+        return True
+    
     def get_download_progress(self):
         ret = {
             "progress": (self.state.download_bytes_now / self.state.download_bytes_total)*100,
@@ -172,8 +200,56 @@ class Reflash:
             "start_time": self.state.download_start_time,
             "filename": self.state.download_filename,
         }
-        if self.state.download_state == "FINISHED":
+        if self.state.download_state in ["FINISHED", "CANCELLED"]:
             self.state.download_state = "IDLE"
+            self.state.save()
+        return ret
+
+    def upload_start(self, filename, size, start_time):
+        if self.get_state() != "IDLE":
+            return False
+        if not filename.endswith(".img.xz"):
+            return False
+        path = self.images_folder+"/"+filename
+        with open(path, 'ab+'):
+            os.utime(path)
+        self.state.upload_state = "UPLOADING"
+        self.state.upload_filename = filename
+        self.state.upload_bytes_total = size
+        self.state.upload_bytes_now = 0
+        self.state.upload_start_time = start_time
+        self.state.save()
+        return True
+
+    def upload_finish(self):
+        self.state.upload_state = "FINISHED"
+        self.state.save()
+        return True
+
+    def upload_cancel(self):
+        self.state.upload_state = "CANCELLED"
+        self.state.save()
+        return True
+
+    def upload_chunk(self, chunk):
+        path = self.images_folder+"/"+self.state.upload_filename
+        with open(path, "ab+") as f:
+            f.write(chunk)
+        self.state.upload_bytes_now += len(chunk)
+        self.state.save('upload_bytes_now')
+        return True
+
+    def get_upload_progress(self):
+        ret = {
+            "state": self.state.upload_state,
+            "progress": (self.state.upload_bytes_now/self.state.upload_bytes_total)*100,
+            "filename": self.state.upload_filename,
+            "bytes_total": self.state.upload_bytes_total,
+            "bytes_now": self.state.upload_bytes_now,
+            "start_time": self.state.upload_start_time,
+        }
+        if self.state.upload_state not in ["UPLOADING", "IDLE"]:
+            self.state.upload_state = "IDLE"
             self.state.save()
         return ret
 
@@ -191,36 +267,23 @@ class Reflash:
         self.state.install_state = "INSTALLING"
         self.state.install_cancelled = False
         self.state.save()
-        ex = self.executor.submit(self.ex_install_refactor, infile)
+        self.executor.submit(self._ex_install_refactor, infile)
         return True
 
-    def ex_install_refactor(self, filename):
-        cmd = ["sudo", "/usr/local/bin/flash-recore", filename]
-        self.process = subprocess.Popen(cmd, stdout=subprocess.PIPE, close_fds=True)
-        while True:
-            time.sleep(1)
-            tr = Reflash.run_system_command("tail -1  /tmp/recore-flash-progress")
-            ti = Reflash.run_system_command("cat /tmp/recore-flash-log")
-            try:
-                self.state.install_progress = float(tr.strip())
-                self.state.install_log = ti
-                self.state.save()
-            except ValueError:
-                pass
-            except Exception as err:
-                self.state.install_state = "ERROR"
-                self.state.install_log += "\nError getting progress and log: "+str(err)
-                self.state.save()
-                break
-            if self.process.poll() == 0:
-                self.state.install_state = "FINISHED"
-                self.state.save()
-                break
-            if self.state.install_state == "CANCELLED":
-                self.state.install_log += "\nInstallation cancelled"
-                break
+    def _ex_install_refactor(self, filename):
+        cmd = " ".join([self.sudo, "/usr/local/bin/flash-recore", filename])
+        self._run_system_command(cmd)
+        if self.state.install_state != "CANCELLED":
+            self.state.install_state = "FINISHED"
+            self.state.save()
 
     def get_install_progress(self):
+        tr = self._run_system_command("tail -1 /tmp/recore-flash-progress")
+        ti = self._run_system_command("cat /tmp/recore-flash-log")
+        self.state.install_progress = self._parse_float(tr)
+        self.state.install_log = ti
+        self.state.save()
+
         ret = {
             "state": self.state.install_state,
             "progress": self.state.install_progress,
@@ -235,7 +298,7 @@ class Reflash:
         return ret
 
     def cancel_installation(self):
-        Reflash.run_system_command("sudo pkill -f xz -9")
+        self._run_system_command(self.sudo+" pkill -f xz -9")
         self.state.install_state = "CANCELLED"
         self.state.save()
         return True
@@ -248,10 +311,30 @@ class Reflash:
         self.state.backup_log = ""
         self.state.backup_start_time = start_time
         self.state.save()
-        ex = self.executor.submit(self.ex_backup_refactor, outfile)
+        self.executor.submit(self._ex_backup_refactor, outfile)
         return True
 
+    def _ex_backup_refactor(self, filename):
+        cmd = " ".join([self.sudo, "/usr/local/bin/backup-emmc", filename])
+        self._run_system_command(cmd)
+        if self.state.backup_state != "CANCELLED":
+            self.state.backup_state = "FINISHED"
+            self.state.save()
+
+    def _parse_float(self, val):
+        try:
+            val = float(val)
+        except:
+            val = 0.0
+        return val
+
     def get_backup_progress(self):
+        tr = self._run_system_command("tail -1 /tmp/recore-flash-progress")
+        ti = self._run_system_command("cat /tmp/recore-flash-log")
+        self.state.backup_progress = self._parse_float(tr)
+        self.state.backup_log = ti
+        self.state.save()
+
         ret = {
             "state": self.state.backup_state,
             "progress": self.state.backup_progress,
@@ -265,70 +348,41 @@ class Reflash:
         return ret
 
     def cancel_backup(self):
-        Reflash.run_system_command("sudo pkill -f backup-emmc -9")
+        self._run_system_command(self.sudo+" pkill -f backup-emmc -9")
         self.state.backup_state = "CANCELLED"
         self.state.save()
         return True
 
-    def ex_backup_refactor(self, filename):
-        cmd = ["sudo", "/usr/local/bin/backup-emmc", filename]
-        self.backup_transferred = 0
-        self.backup_process = subprocess.Popen(cmd, stdout=subprocess.PIPE, close_fds=True)
-        while True:
-            time.sleep(0.3)
-            status = self.backup_process.poll()
-            tr = Reflash.run_system_command("tail -1 /tmp/recore-flash-progress")
-            ti = Reflash.run_system_command("cat /tmp/recore-flash-log")
-            try:
-                self.state.backup_progress = float(tr.strip())
-                self.state.backup_log = ti
-                self.state.save()
-            except:
-                pass
-            if status == 0:
-                self.state.backup_state = "FINISHED"
-                self.state.save()
-                break
-            if status != None:
-                self.state.backup_state = "ERROR"
-                self.state.backup_log += f"\nError {status}"
-                self.state.save()
-                break
-            if self.state.backup_state == "CANCELLED":
-                break
-            self.state.save()
+    def save_options(self, options):
+        if 'darkmode' in options:
+            self.state.settings_darkmode = options['darkmode']
+        if 'rebootWhenDone' in options:
+            self.state.settings_reboot_when_done = options['rebootWhenDone']
+        if 'enableSsh' in options:
+            self.state.settings_enable_ssh = options['enableSsh']
+        self.state.save()
+        return True
 
-    def run_system_command(command):
+    def get_options(self):
+        return {
+            'darkmode': self.state.settings_darkmode,
+            'rebootWhenDone': self.state.settings_reboot_when_done,
+            'enableSsh': self.state.settings_enable_ssh,
+        }
+
+    def _run_system_command(self, command):
         return subprocess.run(command.split(),
                               capture_output=True,
                               text=True).stdout.strip()
 
-    def save_settings(self, settings):
-        import json
-        json_file = json.dumps(settings, indent=4)
-        with open(self.settings_folder+"/settings.json", "w+") as f:
-            f.write(json_file)
-        return True
-
-    def read_settings(self):
-        import json
-        path = self.settings_folder+"/settings.json"
-        if not os.path.isfile(path):
-            return {}
-        with open(path, "r") as f:
-            try:
-                return json.load(f)
-            except:
-                return {}
-
     def enable_ssh(self):
-        return Reflash.run_system_command("sudo /usr/local/bin/enable-emmc-ssh")
+        return self._run_system_command(self.sudo+" /usr/local/bin/enable-emmc-ssh")
 
     def reboot(self):
-        return Reflash.run_system_command("sudo /usr/local/bin/reboot-board")
+        return self._run_system_command(self.sudo+" /usr/local/bin/reboot-board")
 
     def shutdown(self):
-        return Reflash.run_system_command("sudo /usr/local/bin/shutdown-board")
+        return self._run_system_command(self.sudo+" /usr/local/bin/shutdown-board")
 
     def set_boot_media(self, media):
-        return os.system(f"sudo /usr/local/bin/set-boot-media {media}")
+        return self._run_system_command(f"{self.sudo} /usr/local/bin/set-boot-media {media}")
